sandbox: Move the JdbcLedgerDao owner constructor to its object.
